+++ components/org.wso2.carbon.identity.api.user.approval/org.wso2.carbon.identity.api.user.approval.common/src/main/java/org/wso2/carbon/identity/api/user/approval/common/ApprovalConstant.java
@@ -25,6 +25,7 @@ public class ApprovalConstant {
     public static final String USER_APPROVAL_TASK_PATH_COMPONENT = "/%s/approval-tasks";
     public static final String V1_API_PATH_COMPONENT = "/v1";
     public static final String ME_CONTEXT = "me";
+    public static final String SIMPLE_WORKFLOW_ENGINE = "Workflow.SimpleWorkflow.Enable";
 
     /**
      * Enum for user's pending approval related errors in the format of
